optimization changes

+++ clm/SevenpointLevenbergMarquardt.cpp
@@ -22,7 +22,7 @@ void SevenpointLevenbergMarquardtInit() {
     source_file("clm.m");
 }
 
-bool SevenpointLevenbergMarquardt(std::vector<double>* F, std::vector<double> x1, std::vector<double> y1, std::vector<double> x2, std::vector<double> y2, int F0, int maxIter, double stopDist)
+bool SevenpointLevenbergMarquardt(std::vector<double>* F, std::vector<double> x1, std::vector<double> y1, std::vector<double> x2, std::vector<double> y2, double F0, int maxIter, double stopDist)
 {
 
     Matrix X1o = Matrix(x1.size(),1);
