+++ clm/SevenpointLevenbergMarquardt.h
@@ -5,6 +5,6 @@
 
 void SevenpointLevenbergMarquardtExit();
 void SevenpointLevenbergMarquardtInit();
-bool SevenpointLevenbergMarquardt(std::vector<double>  *F, std::vector<double> x1, std::vector<double> y1, std::vector<double> x2, std::vector<double> y2, int F0, int maxIter, double stopDist);
+bool SevenpointLevenbergMarquardt(std::vector<double>  *F, std::vector<double> x1, std::vector<double> y1, std::vector<double> x2, std::vector<double> y2, double F0, int maxIter, double stopDist);
 
 #endif // SEVENPOINTLEVENBERGMARQUARDT_H
